review changes

+++ src/cypher/IRateLimitedAllowance.sol
@@ -5,7 +5,7 @@ interface IRateLimitedAllowance {
     function approve(
         address token,
         address spender,
-        uint160 amount,
+        uint256 amount,
         uint128 rateLimitPerSecond,
         uint128 bufferCap
     ) external;
@@ -13,7 +13,7 @@ interface IRateLimitedAllowance {
     function transferFrom(
         address from,
         address to,
-        uint160 amount,
+        uint256 amount,
         address token
     ) external;
 }
